requirements.txt
flask==3.0.0
playwright==1.40.0
anthropic==0.39.0
requests==2.31.0
PyPDF2==3.0.1
pandas==2.1.4
numpy==1.26.2
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
gunicorn==21.2.0
matplotlib==3.8.2